Revert "Bld 3686"
