+++ Conversation-Samples/HistorySample/Podfile
@@ -8,29 +8,25 @@ platform :ios, "9.0"
 use_frameworks!
 
 # Uncomment to use a local copy of PlayKit
-BLD_UI="../../../BoldAI/BoldUI/BoldUI-Dev.podspec"
-BLD_AI_ENGINE="../../../BoldAI/BoldAIEngine/BoldAIEngine-Dev.podspec"
-BLD_AI_ACCESSIBILITY="../../../BoldAI/BoldAIAccessibility/BoldAIAccessibility-Dev.podspec"
-BLD_ENGINE="../../../BoldEngine/BoldEngine-Dev.podspec"
-BLD_360_AI="../../../Bold360AI/Bold360AI-Dev.podspec"
+# NR_LOCAL="../../../NanorepUI/NanorepUI-Dev.podspec"
+NR_ENGINE="../../../NanorepEngine/NanorepEngine-Dev.podspec"
+NR_ACCESSIBILITY="../../../NRAccessibility/NRAccessibility-Dev.podspec"
 
 # Dev
-source 'https://github.com/nanorepsdk/NRSDK-Dev-specs.git'
+#source 'https://github.com/nanorepsdk/NRSDK-Dev-specs.git'
 # Production
-#source 'https://github.com/nanorepsdk/NRSDK-specs.git'
+source 'https://github.com/nanorepsdk/NRSDK-specs.git'
 
 source 'https://github.com/CocoaPods/Specs'
 
 target 'HistorySample' do
-    if defined? BLD_UI 
-        pod 'BoldAIEngine', :path => BLD_AI_ENGINE
-        pod 'BoldAIAccessibility', :path => BLD_AI_ACCESSIBILITY
-        pod 'BoldEngine', :path => BLD_ENGINE
-        pod 'BoldUI', :path => BLD_UI
-     #   pod 'Bold360AI', :path => BLD_360_AI
+    if defined? NR_LOCAL 
+        pod 'NanorepEngine', :path => NR_ENGINE
+        pod 'NRAccessibility', :path => NR_ACCESSIBILITY
+        pod 'NanorepUI', :path => NR_LOCAL
         pod 'RealmSwift'
     else
-        pod 'Bold360AI'
+        pod 'NanorepUI'
         pod 'RealmSwift'
     end
 end
